+++ src/rend/color/mod.rs
@@ -3,13 +3,14 @@
 //! Chromatic functionality.
 //
 
+mod error;
 mod r#struct;
 mod r#trait;
 #[allow(unused_imports)]
-pub use {r#struct::*, r#trait::*};
+pub use {error::*, r#struct::*, r#trait::*};
 
 pub(crate) mod all {
     #[doc(inline)]
     #[allow(unused_imports)]
-    pub use super::{r#struct::*, r#trait::*};
+    pub use super::{error::*, r#struct::*, r#trait::*};
 }